+++ R/conv.R
@@ -13,5 +13,9 @@
 #' str_conv(x, "ISO-8859-2") # Polish "a with ogonek"
 #' str_conv(x, "ISO-8859-1") # Plus-minus
 str_conv <- function(string, encoding) {
+  if (!is_string(encoding)) {
+    abort("`encoding` must be a single string.")
+  }
+
   stri_conv(string, encoding, "UTF-8")
 }