+++ R/locate.r
@@ -26,8 +26,10 @@
 #' # Find location of every character
 #' str_locate_all(fruit, "")
 str_locate <- function(string, pattern) {
+  check_lengths(string, pattern)
+
   switch(type(pattern),
-    empty = stri_locate_first_boundaries(string, opts_brkiter = opts(pattern)),
+    empty = ,
     bound = stri_locate_first_boundaries(string, opts_brkiter = opts(pattern)),
     fixed = stri_locate_first_fixed(string, pattern, opts_fixed = opts(pattern)),
     coll  = stri_locate_first_coll(string, pattern, opts_collator = opts(pattern)),
@@ -38,10 +40,11 @@ str_locate <- function(string, pattern) {
 #' @rdname str_locate
 #' @export
 str_locate_all <- function(string, pattern) {
+  check_lengths(string, pattern)
   opts <- opts(pattern)
 
   switch(type(pattern),
-    empty = stri_locate_all_boundaries(string, omit_no_match = TRUE, opts_brkiter = opts),
+    empty = ,
     bound = stri_locate_all_boundaries(string, omit_no_match = TRUE, opts_brkiter = opts),
     fixed = stri_locate_all_fixed(string, pattern, omit_no_match = TRUE, opts_fixed = opts),
     regex = stri_locate_all_regex(string, pattern, omit_no_match = TRUE, opts_regex = opts),