+++ R/utils.R
@@ -7,3 +7,14 @@
 #' @importFrom magrittr %>%
 #' @usage lhs \%>\% rhs
 NULL
+
+check_lengths <- function(string, pattern, replacement = NULL) {
+  # stringi already correctly recycles vectors of length 0 and 1
+  # we just want more stringent vctrs checks for other lengths
+  vctrs::vec_size_common(
+    string = string,
+    pattern = pattern,
+    replacement = replacement
+  )
+}
+